Rename folder Text to Paragraph
