+++ packages/react/src/components/Accordion/AccordionContent.tsx
@@ -2,7 +2,7 @@ import cl from 'classnames';
 import React, { forwardRef, useContext } from 'react';
 
 import AnimateHeight from '../../utils/AnimateHeight';
-import { Text } from '../Typography/Text';
+import { Text } from '../Typography/Paragraph';
 
 import classes from './Accordion.module.css';
 import { AccordionItemContext } from './AccordionItem';
