+++ packages/react/src/components/Typography/index.ts
@@ -1,2 +1,2 @@
-export * from './Text';
+export * from './Paragraph';
 export * from './Heading';
